Create 0000-TheAlgorithms.org and 0001-The-Structures.org
The former file is for looking at algorithms (this is not exclusive to
just sorting algorithms). The latter file is for a more atomic
approach in the data structures of which the algorithm may be used.

Note that the number is made just for convention, but does not
necessarily dictate the order of which these files should be read